refactor(proxy): make hardware.cjs the source of truth, hardware.js an ESM wrapper (refs #324, Strategy C)
Resolves the ESM/CJS boundary so Electron's CJS main process can
require('@xiboplayer/proxy/hardware') and get the same logic the
Chromium kiosk + ESM consumers use. One source of truth.

Implementation moves:
- hardware.cjs: full 150-LOC implementation in CommonJS form
  (module.exports). Previously a 'not implemented' stub. Pure Node-
  core (fs, path, os) — no runtime dependencies on ESM-only features.
- hardware.js: shrinks to 12 LOC of `createRequire` + re-exports.
  Same named exports (GPU_VENDORS, detectGPUs, selectGPU,
  getMemoryTuning, getHardwareConfig), identical call signatures,
  identical return shapes.

Existing ESM importers are unchanged:
- packages/proxy/src/proxy.js          → import { getHardwareConfig }
- packages/proxy/src/index.js          → re-exports all
- packages/proxy/bin/detect-hardware.js → import { getHardwareConfig }

They now flow through the ESM wrapper → createRequire → CJS impl.
Smoke-tested: both entry points return identical results
(13 flags, same memory tuning, same GPU_VENDORS table).

Electron-side migration (separate PR on xiboplayer-electron repo):
drop the duplicate GPU_VENDORS + detectGPUs() copy from
xiboplayer-electron/src/main.js:81-100+, replace with

    const { getHardwareConfig } = require('@xiboplayer/proxy/hardware');

Net effect: ~200 LOC removed from the Electron wrapper, single source
for future flag additions (e.g. adding a new Chromium command-line
switch only needs to land here).
